requirements.txt
beautifulsoup4==4.12.3
click==8.0.3
dynaconf==3.1.7
httpx==0.27.2
loguru==0.6.0
lxml_html_clean==0.2.1
pybit7z==0.3.2
pyppeteer==2.0.0
PyYAML==6.0.2
requests==2.32.3
requests-html==0.10.0
setuptools==75.6.0
